convert tooltip mixin into a custom directive to simplify tooltip setup
