+++ pgmanage/app/static/pgmanage_frontend/src/components/RestoreTab.vue
@@ -25,6 +25,7 @@
 
               <div
                 v-if="!isNotServer && pgKeys.length"
+                v-tooltip
                 class="form-group d-flex flex-column justify-content-end"
                 data-bs-toggle="tooltip"
                 data-bs-placement="bottom"
@@ -73,6 +74,7 @@
               <div v-if="isNotServer" class="row mt-1">
                 <div
                   v-if="pgKeys.length"
+                  v-tooltip
                   class="form-group col-4 d-flex flex-column justify-content-end"
                   data-bs-toggle="tooltip"
                   data-bs-placement="bottom"
@@ -305,7 +307,6 @@ import axios from 'axios'
 import { showAlert } from '../notification_control';
 import { settingsStore, fileManagerStore, tabsStore } from '../stores/stores_initializer';
 import { truncateText } from "../utils";
-import ComponentTooltipMixin from "../mixins/component_tooltip_mixin";
 import { handleError } from '../logging/utils';
 
 export default {
@@ -313,7 +314,6 @@ export default {
   components: {
     UtilityJobs,
   },
-  mixins: [ComponentTooltipMixin],
   props: {
     workspaceId: String,
     tabId: String,
